irqchip/gic-v3: Add a DMA Non-Coherent flag
The GIC600 integration in RK356x, used in rk3588, doesn't support
any of the shareability or cacheability attributes, and requires
both values to be set to 0b00 for all the ITS and Redistributor
tables.

This is loosely based on prior work from XiaoDong Huang and
Peter Geis fixing this issue specifically for Rockchip 356x.

Suggested-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Lucas Tanure <lucas.tanure@collabora.com>
Change-Id: I4821086a9723bbe163b81775b85120aefc8d69d5