[BROKEN] arm64: dts: rockchip: rk3588-rock5b: add USB-C support
Add support for using the Radxa Rock 5 Model B USB-C port for USB in
OHCI, EHCI or XHCI mode. Displayport AltMode is not yet supported.

Note: Enabling support for the USB-C port results in a board reset
when the system is supplied with a PD capable power-supply. Until
this has been analyzed and fixed, let's disable support for the
Type-C port.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>