[BROKEN] pmdomain: rockchip: allow building as module
Now that power-domain support is initialized using device_initcall,
let's go one step further and allow building it as a module.

NOTE: This allows building as a module, but it does not really
work at runtime. The driver is probed very late resulting in a
bunch of other drivers ending up with -ETIMEDOUT before the kernel
probed pm-domain support. This results in the boot devices never
appearing and a boot failure.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>