drm/rockchip: dw_hdmi_qp: Make use of phy_configure()
Stop relying on phy_set_bus_width() workaround to setup the TMDS
character rate and, instead, use the recently introduced HDMI PHY
configuration API.

Additionally, move the logic to ->atomic_check() callback where the
current mode rate is already provided by the connector state.  As a
matter of fact this is actually necessary to ensure the rate is
configured before VOP2 attempts to use the PHY PLL as a DCLK source in
vop2_crtc_atomic_enable().  The rationale is to restrict any changes of
the PHY rate via CCF and, instead, prefer the PHY configuration API for
this purpose.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>