[NOUPSTREAM] gitlab-ci: enable a few more filesystems
The arm64 defconfig does not enable all filesystems one might
reasonably expect. While it's fine not to enable everything under
the sun, some of these are actually useful in an embedded use-case
and reasonably maintained.

Enable F2FS, EROFS, ISO9660, EXFAT and XFS for the benefit of both
casual users and people setting up demos on the board.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>