PCI: Restore assigned resources fully after release
PCI resource fitting code in __assign_resources_sorted() runs in
multiple steps. A resource that was successfully assigned may have to
be released before the next step attempts assignment again. The
assign+release cycle is destructive to a start-aligned struct resource
(bridge window or IOV resource) because the start field is overwritten
with the real address when the resource got assigned.

Properly restore the resource after releasing it. The start, end, and
flags fields must be stored into the related struct pci_dev_resource in
order to be able to restore the resource to its original state.

Reported-by: Guenter Roeck <linux@roeck-us.net>
Fixes: 96336ec70264 ("PCI: Perform reset_resource() and build fail list in sync")
Signed-off-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>
Tested-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>
Tested-by: Ondrej Jirman <megi@xff.cz>
Tested-by: Guenter Roeck <linux@roeck-us.net>