[HACK] arm64: dts: rockchip: rock5bp: synchronize HDMI ports
Synchronize both HDMI outputs. This is a big hack and not meant for
general upstream support. It requires, that both ports are connected
to equal displays with the goal that both of them get their vsync
at exactly the same time. This can be used to experiment with VR-like
setups with one display per eye.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>