[HACK] arm64: dts: rockchip: make npu regulator on ROCK 5B always-on
While having the NPU regulator as a domain-supply for PD_NPU did work
wonderfully for me at one point, it somehow stopped working now.
Specifically, the nputop PD throws an SError while trying to set itself
to idle, which is curious as it's a child domain of PD_NPU.

Mark vdd_npu_s0 as always-on for now until I can take the biggest
sledgehammer imaginable to the Rockchip genpd driver, for emotional
relief if not for technical reasons. Something seems genuinely busted
and I'm not sure why we go through the trouble of even manually
specifying whether each PD "wants" a regulator in that driver when what
it does in reality is just ignore any domain-supply unless the PD
"wants" a regulator, thereby just acting as a mask for the DT.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>