drm/rockchip: dw_hdmi_qp: Fixup usage of enable_gpio member in main struct
The name of the enable_gpio member in struct rockchip_hdmi_qp is too
generic, it was "borrowed" from downstream BSP code.  Moreover, this
hasn't been really in use so far, since there is neither a DT providing
an "enable-gpios" property to any of the "hdmi" nodes, nor a binding
documenting it.

The actual purpose of this GPIO line is to control the level shifters
for the HDMI TX data lines when switching between TMDS and FRL operating
modes, i.e. asserted for the former and deasserted for the latter.

This has been recently documented via the "tmds-enable-gpios" property
of the RK3588 HDMI TX Controller binding, hence pass the correct string
to devm_gpiod_get_optional() and rename the struct member accordingly.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>