drm/rockchip: dw_hdmi_qp: Improve error handling with dev_err_probe()
The error handling in dw_hdmi_qp_rockchip_bind() is quite inconsistent,
i.e. in some cases the error code is not included in the message, while
in some other cases there is no check for -EPROBE_DEFER.

Since this is part of the probe path, address the aforementioned issues
by switching to dev_err_probe(), which also reduces the code a bit.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>
Link: https://lore.kernel.org/r/20250710-rk3588-hdmi-cec-v2-3-f5884be34bc1@collabora.com
Signed-off-by: Sebastian Reichel <sre@kernel.org>