arm64: dts: rockchip: Fix USB-C description for RK3588 EVB1
Fix the USB-C connector description, so that it follows the binding:

port@0 is the high-speed lanes and should point to the USB2 PHY
port@1 is the super-speed lanes and should point to the USB3 PHY
port@2 is the SBU lanes and should point to the DP PHY

Right now the high-speed and super-speed links are swapped and
for the high-speed lanes the link incorrectly points to the
controller instead of the PHY.

This also updates the port naming, so that it describes the hardware
instead of how the drivers are using the information. These are
effectively the same, but the DT should describe hardware and not
software.

Fixes: b37146b5a555 ("arm64: dts: rockchip: add USB3 to rk3588-evb1")
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>